// src/generated/schema-types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
    ID: { input: string; output: string; }
    String: { input: string; output: string; }
    Boolean: { input: boolean; output: boolean; }
    Int: { input: number; output: number; }
    Float: { input: number; output: number; }
    CompareValue: { input: any; output: any; }
    /** A date. */
    Date: { input: any; output: any; }
    /** A multipart file */
    File: { input: any; output: any; }
    /** An ISO 8601-encoded datetime */
    ISO8601DateTime: { input: any; output: any; }
    /** A JSON formatted string. */
    JSON: { input: any; output: any; }
};
// ...
export type Column = {
    __typename?: 'Column';
    /** Is the column archived or not. */
    archived: Scalars['Boolean']['output'];
    /** The column's description. */
    description?: Maybe<Scalars['String']['output']>;
    /** The column's unique identifier. */
    id: Scalars['ID']['output'];
    /** The column's settings in a string form. */
    settings_str: Scalars['String']['output'];
    /** The column's title. */
    title: Scalars['String']['output'];
    /** The column's type. */
    type: ColumnType;
    /** The column's width. */
    width?: Maybe<Scalars['Int']['output']>;
};
// ...
export enum ColumnType {
    /** Number items according to their order in the group/board */
    AutoNumber = 'auto_number',
    /** Connect data from other boards */
    BoardRelation = 'board_relation',
    /** Perform actions on items by clicking a button */
    Button = 'button',
    /** Check off items and see what's done at a glance */
    Checkbox = 'checkbox',
    /** Manage a design system using a color palette */
    ColorPicker = 'color_picker',
    /** Choose a country */
    Country = 'country',
    /** Add the item's creator and creation date automatically */
    CreationLog = 'creation_log',
    /** Add dates like deadlines to ensure you never drop the ball */
    Date = 'date',
    /** Set up dependencies between items in the board */
    Dependency = 'dependency',
    /** Instantly add collaborative rich text editor */
    Doc = 'doc',
    /** Create a dropdown list of options */
    Dropdown = 'dropdown',
    /** Email team members and clients directly from your board */
    Email = 'email',
    /** Add files & docs to your item */
    File = 'file',
    /** Use functions to manipulate data across multiple columns */
    Formula = 'formula',
    Group = 'group',
    /** Add times to manage and schedule tasks, shifts and more */
    Hour = 'hour',
    /** Integration is really cool... */
    Integration = 'integration',
    /** Show all item's assignees */
    ItemAssignees = 'item_assignees',
    /** Show a unique ID for each item */
    ItemId = 'item_id',
    /** Add the person that last updated the item and the date */
    LastUpdated = 'last_updated',
    /** Simply hyperlink to any website */
    Link = 'link',
    /** Place multiple locations on a geographic map */
    Location = 'location',
    /** Add large amounts of text without changing column width */
    LongText = 'long_text',
    /** Show and edit columns' data from connected boards */
    Mirror = 'mirror',
    /** Name is really cool... */
    Name = 'name',
    /** Add revenue, costs, time estimations and more */
    Numbers = 'numbers',
    /** Assign people to improve team work */
    People = 'people',
    /** Assign a person to increase ownership and accountability (deprecated) */
    Person = 'person',
    /** Call your contacts directly from monday.com */
    Phone = 'phone',
    /** Show progress by combining status columns in a battery */
    Progress = 'progress',
    /** Rate or rank anything visually */
    Rating = 'rating',
    /** Get an instant overview of where things stand */
    Status = 'status',
    /** Use the subtasks column to create another level of tasks */
    Subtasks = 'subtasks',
    /** Add tags to categorize items across multiple boards */
    Tags = 'tags',
    /** Assign a full team to an item */
    Team = 'team',
    /** Add textual information e.g. addresses, names or keywords */
    Text = 'text',
    /** Easily track time spent on each item, group, and board */
    TimeTracking = 'time_tracking',
    /** Visualize your item’s duration, with a start and end date */
    Timeline = 'timeline',
    /** Unsupported column type */
    Unsupported = 'unsupported',
    /** Vote on an item e.g. pick a new feature or a favorite lunch place */
    Vote = 'vote',
    /** Select the week on which each item should be completed */
    Week = 'week',
    /** Keep track of the time anywhere in the world */
    WorldClock = 'world_clock'
}
// ...
export type Country = {
    __typename?: 'Country';
    /** The country's two-letter code. */
    code: Scalars['String']['output'];
    /** The country's name. */
    name: Scalars['String']['output'];
};
// ...
export type Group = {
    __typename?: 'Group';
    /** Is the group archived or not. */
    archived?: Maybe<Scalars['Boolean']['output']>;
    /** The group's color. */
    color: Scalars['String']['output'];
    /** Is the group deleted or not. */
    deleted?: Maybe<Scalars['Boolean']['output']>;
    /** The group's unique identifier. */
    id: Scalars['ID']['output'];
    /** The items in the group. */
    items_page: ItemsResponse;
    /** The group's position in the board. */
    position: Scalars['String']['output'];
    /** The group's title. */
    title: Scalars['String']['output'];
};
// ...
export type ItemsResponse = {
    __typename?: 'ItemsResponse';
    /**
     * An opaque cursor that represents the position in the list after the last
     * returned item. Use this cursor for pagination to fetch the next set of items.
     * If the cursor is null, there are no more items to fetch.
     */
    cursor?: Maybe<Scalars['String']['output']>;
    /** The items associated with the cursor. */
    items: Array<Item>;
};
// ...
export type Team = {
    __typename?: 'Team';
    /** The team's unique identifier. */
    id: Scalars['ID']['output'];
    /** The team's name. */
    name: Scalars['String']['output'];
    /** The users who are the owners of the team. */
    owners: Array<User>;
    /** The team's picture url. */
    picture_url?: Maybe<Scalars['String']['output']>;
    /** The users in the team. */
    users?: Maybe<Array<Maybe<User>>>;
};
// ...
export type User = {
    __typename?: 'User';
    /** The user's account. */
    account: Account;
    /** The user's birthday. */
    birthday?: Maybe<Scalars['Date']['output']>;
    /** The user's country code. */
    country_code?: Maybe<Scalars['String']['output']>;
    /** The user's creation date. */
    created_at?: Maybe<Scalars['Date']['output']>;
    /** The current user's language */
    current_language?: Maybe<Scalars['String']['output']>;
    /** The user's email. */
    email: Scalars['String']['output'];
    /** Is the user enabled or not. */
    enabled: Scalars['Boolean']['output'];
    /** The user's unique identifier. */
    id: Scalars['ID']['output'];
    /** Is the user an account admin. */
    is_admin?: Maybe<Scalars['Boolean']['output']>;
    /** Is the user a guest or not. */
    is_guest?: Maybe<Scalars['Boolean']['output']>;
    /** Is the user a pending user */
    is_pending?: Maybe<Scalars['Boolean']['output']>;
    /** Is user verified his email. */
    is_verified?: Maybe<Scalars['Boolean']['output']>;
    /** Is the user a view only user or not. */
    is_view_only?: Maybe<Scalars['Boolean']['output']>;
    /** The date the user joined the account. */
    join_date?: Maybe<Scalars['Date']['output']>;
    /** Last date & time when user was active */
    last_activity?: Maybe<Scalars['Date']['output']>;
    /** The user's location. */
    location?: Maybe<Scalars['String']['output']>;
    /** The user's mobile phone number. */
    mobile_phone?: Maybe<Scalars['String']['output']>;
    /** The user's name. */
    name: Scalars['String']['output'];
    /** The user's out of office status. */
    out_of_office?: Maybe<OutOfOffice>;
    /** The user's phone number. */
    phone?: Maybe<Scalars['String']['output']>;
    /** The user's photo in the original size. */
    photo_original?: Maybe<Scalars['String']['output']>;
    /** The user's photo in small size (150x150). */
    photo_small?: Maybe<Scalars['String']['output']>;
    /** The user's photo in thumbnail size (100x100). */
    photo_thumb?: Maybe<Scalars['String']['output']>;
    /** The user's photo in small thumbnail size (50x50). */
    photo_thumb_small?: Maybe<Scalars['String']['output']>;
    /** The user's photo in tiny size (30x30). */
    photo_tiny?: Maybe<Scalars['String']['output']>;
    /** The product to which the user signed up to first. */
    sign_up_product_kind?: Maybe<Scalars['String']['output']>;
    /** The teams the user is a member in. */
    teams?: Maybe<Array<Maybe<Team>>>;
    /** The user's timezone identifier. */
    time_zone_identifier?: Maybe<Scalars['String']['output']>;
    /** The user's title. */
    title?: Maybe<Scalars['String']['output']>;
    /** The user's profile url. */
    url: Scalars['String']['output'];
    /** The user’s utc hours difference. */
    utc_hours_diff?: Maybe<Scalars['Int']['output']>;
};